backend/requirement.txt
langchain
python-dotenv
langchain-community
tiktoken
faiss-cpu
langchain-openai
pypdf
beautifulsoup4
validators
selenium
webdriver_manager
pymysql